OM5K-7987 - Adding new report config force user to key in email address
Error was caused by the input.replace throwing an error as input was undefined due to no user input thus the replace properly was missing which meant the email_array split also returned an error failing the loop. The entire function was wrapped in a promise so the entire function threw a failed state causing the odd Error State for that form.