OM5K-9886 Add a view-only column “% or PPM” in ratios table in GUI. It can automatically calculate and display the percentage. It will show both value and tip text in the grid cell.
The logic used is as follows:

If the base product in recipe has the configuration in pipenode, we will use the pipenode to identify the relationship between the base product and the additives, and use it to calculate the percentage. For example, if recipe has B1, B2, A1, A2, A3, and in pipenode A1 and A2 are configured to inject to B1, A3 configured to B2, the percentage of B1 will be (B1+A1+A2)/(B1+A1+A2+B2+A3), and the percentage of B2 will be (B2+A3)/(B1+A1+A2+B2+A3)
If the base product in recipe has not been configured in pipenode yet, we will assume all additives belong to the base product with highest ratio, and use it to calculate the percentage. For example, if recipe has B1, B2, A1, A2, A3, and B1>B2>A1,A2,A3, the percentage of B1 will be (B1+A1+A2+A3)/(B1+A1+A2+B2+A3), and the percentage of B2 will be B2/(B1+A1+A2+B2+A3)